test(task): split test_state_machine.py + tighten BUG-TSK-001 assertion (Norman R-N1)
Norman R-N1 [却下] への対応。Track A workflow-repository PR #41 の 0.4%超過
を却下した直後に本 PR で 27% 超過 (633行) を通すと一貫性原理が壊れるため、
3 ファイル分割 + BUG-TSK-001 修正後の assertion 収斂を実施。

①ファイル分割 (633→3 files、500 行ルール準拠):

* `_helpers.py` (115 行) — 5 ヘルパー (`ALL_ACTIONS` / `ALL_STATUSES` /
  `next_ts` / `invoke_action` / `make_task_in_status`) を package-private
  module に抽出 (underscore prefix で pytest 収集対象から除外)
* `test_state_machine_table.py` (177 行) — TestStateMachineTableLocked
  + TestThirteenAllowedTransitions (TC-UT-TS-039 + 13 ✓ cells)
* `test_state_terminal_and_invalid.py` (96 行) — TestTerminalGate
  (20 ✗ DONE/CANCELLED) + TestEveryIllegalNonTerminalCellRaises
  (27 ✗ non-terminal)
* `test_state_lifecycle.py` (335 行) — TestBlockRequiresNonEmptyLastError
  + TestPreValidateLeavesOriginalUntouched + TestLifecycleIntegration
  + TestUpdatedAtAdvances

②BUG-TSK-001 assertion 収斂 (regression-blind 排除):

実装は commit 377366e で `_check_invariants` の validator 順序を入れ替え、
`block(last_error='')` が `blocked_requires_last_error` (MSG-TS-006) を
emit する状態に収斂済み。test 側も `assert kind in {"blocked_requires_last_error",
"last_error_consistency"}` の二択許容を `assert kind == "blocked_requires_last_error"`
の単一 kind 締めに変更。

* test docstring の "BUG-TSK-001 [MEDIUM]: ..." 記述削除、代わりに「修正
  済み (commit 377366e)」+「regression が validator 順序を逆転させたら
  ここで surface する」を凍結
* test method 名を `test_block_with_empty_last_error_raises` →
  `test_block_with_empty_last_error_raises_blocked_requires_last_error`
  に締める (期待 kind を method 名で表明)
* assertion failure 時の Next: hint で「BUG-TSK-001 fix (commit 377366e)
  を確認しろ」を案内

ファイル一覧 (test_task/ 配下、全て 500 行未満):
* __init__.py                             17
* _helpers.py                            115 (新規、pkg-private helpers)
* test_construction.py                   265
* test_invariants.py                     396
* test_state_lifecycle.py                335 (新規)
* test_state_machine_table.py            177 (新規)
* test_state_terminal_and_invalid.py      96 (新規)
* test_vo.py                             332

品質: ruff clean、pyright strict 0/0、pytest 758 passed (task domain
204 件 + 既存全テスト無回帰)。
